Refactor issue reporter repository for interface-based design
